+++ multimodalhugs/processors/legacy/pose2text_preprocessor.py
@@ -4,6 +4,7 @@ from typing import Any, Optional
 from multimodalhugs.processors.meta_processor import MultimodalMetaProcessor, ProcessorSlot
 from multimodalhugs.processors.pose_modality_processor import PoseModalityProcessor
 from multimodalhugs.processors.text_modality_processor import TextModalityProcessor, TextRole
+from multimodalhugs.processors.utils import SignalUnit
 
 logger = logging.getLogger(__name__)
 
@@ -24,10 +25,12 @@ class Pose2TextTranslationProcessor(MultimodalMetaProcessor):
         tokenizer: Optional[Any] = None,
         reduce_holistic_poses: bool = True,
         skip_frames_stride: Optional[int] = None,
+        signal_start_end_unit: SignalUnit = SignalUnit.MILLISECONDS,
         **kwargs,
     ):
         self.reduce_holistic_poses = reduce_holistic_poses
         self.skip_frames_stride = skip_frames_stride
+        self.signal_start_end_unit = signal_start_end_unit
         # Pass-through for from_pretrained, which calls cls(slots=..., tokenizer=...)
         if "slots" in kwargs:
             super().__init__(tokenizer=tokenizer, **kwargs)
@@ -38,6 +41,7 @@ class Pose2TextTranslationProcessor(MultimodalMetaProcessor):
                     processor=PoseModalityProcessor(
                         reduce_holistic_poses=reduce_holistic_poses,
                         skip_frames_stride=skip_frames_stride,
+                        signal_start_end_unit=signal_start_end_unit,
                     ),
                     output_data_key="input_frames",
                     output_mask_key="attention_mask",