+++ multimodalhugs/processors/legacy/video2text_preprocessor.py
@@ -4,6 +4,7 @@ from typing import Any, Optional
 from multimodalhugs.processors.meta_processor import MultimodalMetaProcessor, ProcessorSlot
 from multimodalhugs.processors.video_modality_processor import VideoModalityProcessor
 from multimodalhugs.processors.text_modality_processor import TextModalityProcessor, TextRole
+from multimodalhugs.processors.utils import SignalUnit
 
 logger = logging.getLogger(__name__)
 
@@ -26,12 +27,14 @@ class Video2TextTranslationProcessor(MultimodalMetaProcessor):
         skip_frames_stride: Optional[int] = None,
         join_chw: bool = False,
         use_cache: bool = True,
+        signal_start_end_unit: SignalUnit = SignalUnit.MILLISECONDS,
         **kwargs,
     ):
         self.custom_preprocessor_path = custom_preprocessor_path
         self.skip_frames_stride = skip_frames_stride
         self.join_chw = join_chw
         self.use_cache = use_cache
+        self.signal_start_end_unit = signal_start_end_unit
         if "slots" in kwargs:
             super().__init__(tokenizer=tokenizer, **kwargs)
             return
@@ -43,6 +46,7 @@ class Video2TextTranslationProcessor(MultimodalMetaProcessor):
                         skip_frames_stride=skip_frames_stride,
                         join_chw=join_chw,
                         use_cache=use_cache,
+                        signal_start_end_unit=signal_start_end_unit,
                     ),
                     output_data_key="input_frames",
                     output_mask_key="attention_mask",